% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data10.R
\docType{data}
\name{example_data10}
\alias{example_data10}
\title{regression_data
This contains estimation and p value data from a/the regression summary output/s.}
\format{
A dataframe with 9 rows and 4 columns
\describe{
\item{x_var}{components displayed in the horizontal axis of the heatmap}
\item{y_var}{components displayed in the vertical axis of the heatmap}
\item{estimate}{regression coefficient}
\item{p_value}{p value}
}
}
\usage{
example_data10
}
\description{
regression_data
This contains estimation and p value data from a/the regression summary output/s.
}
\keyword{datasets}
