% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data11.R
\docType{data}
\name{example_data11}
\alias{example_data11}
\title{CI_data
This contains individual level data for constructing confidence intervals}
\format{
A dataframe with 8 rows and 5 columns
\describe{
\item{IID}{Individual identification number}
\item{PRS}{Estimate (e.g. individual polygenic risk score (PRS))}
\item{Variance}{Variance of the estimate}
\item{Lower_Limit}{Lower limit of the confidence interval}
\item{Upper_Limit}{Upper limit of the confidence interval}
}
}
\usage{
example_data11
}
\description{
CI_data
This contains individual level data for constructing confidence intervals
}
\keyword{datasets}
