% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data2.R
\docType{data}
\name{example_data2}
\alias{example_data2}
\title{gwas_data with partial chromosome information
This contains Single nucleotide polymorphism (SNP) information.}
\format{
A dataframe with 1000 rows and 4 columns
\describe{
\item{SNP}{SNP ID}
\item{CHR}{Chromosome ID}
\item{POS}{Base Pair Position}
\item{P_VALUE}{p value}
}
}
\usage{
example_data2
}
\description{
gwas_data with partial chromosome information
This contains Single nucleotide polymorphism (SNP) information.
}
\keyword{datasets}
