% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data3.R
\docType{data}
\name{example_data3}
\alias{example_data3}
\title{accuracy_data with full information
This contains prediction accuracy related information.}
\format{
A dataframe with 6 rows and 4 columns
\describe{
\item{Model}{Model name}
\item{R_squared}{Measuremnt of prediction accuracy}
\item{p_value}{p value of prediction accuracy}
\item{se}{standard error of prediction accuracy}
}
}
\usage{
example_data3
}
\description{
accuracy_data with full information
This contains prediction accuracy related information.
}
\keyword{datasets}
