% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data4.R
\docType{data}
\name{example_data4}
\alias{example_data4}
\title{accuracy_data with partial information
This contains prediction accuracy related information.}
\format{
A dataframe with 6 rows and 2 columns
\describe{
\item{Model}{Model name}
\item{R_squared}{Measuremnt of prediction accuracy}
}
}
\usage{
example_data4
}
\description{
accuracy_data with partial information
This contains prediction accuracy related information.
}
\keyword{datasets}
