% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data5.R
\docType{data}
\name{example_data5}
\alias{example_data5}
\title{comparison_data
This contains prediction accuracy related information for model comparison.}
\format{
A dataframe with 15 rows and 3 columns
\describe{
\item{Compare1}{Model name 1}
\item{Compare2}{Model name 2}
\item{p_value}{p value of diffence between prediction accuracy}
}
}
\usage{
example_data5
}
\description{
comparison_data
This contains prediction accuracy related information for model comparison.
}
\keyword{datasets}
