% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data7.R
\docType{data}
\name{example_data7}
\alias{example_data7}
\title{ld_data
This contains Linkage Disequilibrium (LD) related data (e.g. pair-wise variant correlation squared).}
\format{
A dataframe with 1000 rows and 7 columns
\describe{
\item{CHR_A}{Chromosome ID}
\item{BP_A}{Base pair position of the 1st variant}
\item{SNP_A}{First variant ID}
\item{CHR_B}{Chromosome ID}
\item{BP_B}{Base pair position of the 2nd variant}
\item{SNP_B}{Second variant ID}
\item{R2}{LD measurement (e.g. R squared)}
}
}
\usage{
example_data7
}
\description{
ld_data
This contains Linkage Disequilibrium (LD) related data (e.g. pair-wise variant correlation squared).
}
\keyword{datasets}
